Ensure boring only compiled when needed

+++ src/core/tsi/ssl_types.h
@@ -41,7 +41,7 @@
 #define TSI_SIZE_AS_SIZE(x) (x)
 #else
 #define TSI_INT_AS_SIZE(x) (x)
-#define TSI_SIZE_AS_SIZE(x) ((size_t)(x))
+#define TSI_SIZE_AS_SIZE(x) ((int)(x))
 #endif
 
 #endif
